CHEAT_007_a Fixed tests for database access [ES,SE]
